+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## NEXT
+
+- Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_adjacency_matrices` for
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not adding floating nodes to
+  the graph. Now any floating nodes at time lag 0 will be added.
+
 ## 0.3.6
 
 - Added `__iter__` to `cai_causal_graph.causal_graph.Skeleton`.